% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinityMatrix.R
\name{affinityMatrix}
\alias{affinityMatrix}
\title{affinityMatrix}
\usage{
affinityMatrix(diff, K = 20, sigma = 0.5)
}
\arguments{
\item{diff}{is a  Distance matrix.
affinityMatrix(diff,K=20,sigma=0.5)}
}
\description{
affinityMatrix
}
