% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_weight.R
\name{cal_weight}
\alias{cal_weight}
\title{calculate weight}
\usage{
cal_weight(Data_list, KNN_list, knn_pred, method = "euclidean")
}
\arguments{
\item{Data_list}{include multi-omics data.
cal_weight(Data_list, KNN_list, knn_pred, method = "euclidean")}
}
\description{
calculate weight
}
