% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_knn.R
\name{generate_knn}
\alias{generate_knn}
\title{Generate Knn}
\usage{
generate_knn(data, row_feature = T, method = "euclidean", K = 20)
}
\arguments{
\item{data}{is a data.frame .
generate_knn(data)}
}
\description{
Generate Knn
}
