% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_prediction.R
\name{knn_prediction}
\alias{knn_prediction}
\title{Knn Prediction}
\usage{
knn_prediction(Data_list, KNN_list, row_feature = T)
}
\arguments{
\item{Data_list}{include multi-omics data.
knn_prediction(Data_list, KNN_list, row_feature = T)}
}
\description{
Knn Prediction
}
